Set profile on all applications now that grail-shell has been restored and requires a profile
